[Section: 2.2] - Fetch Wall Posts on profile view
Description:
Implemented functionality to fetch wall posts on a user's profile view. This allows us to display all the threads/posts made by a particular user on their profile. Leveraged Firestore queries for efficient fetching of data. Also added a pull-to-refresh feature in the feed view to fetch latest threads.

Affected files:
- WallService.swift
- FeedView.swift
- UserContentListViewModel.swift

Notes:

This commit extends the capability to fetch threads for a user profile. Implemented a new function in the thread service to fetch threads based on the user ID. Also, for enhancing user experience, added a pull-to-refresh feature in the main feed view.