[Section 2.3] - Like/unlike wall posts and reply view
Description:

- Liking a wall post available - now it keeps track in our data base for liked posts - also removes -1 for each unlike to keep the accurate amount of likes.
 - rules to not let a user unlike if like count < 0.
- Backend now is storing liked posts as well as which user liked that post.
 - when reseting the app likes will stay in place  since they are being stored in the BE with Firebase Database
- Wall Reply view implemented
- A lot of cleaning was done and re organization of folders.
- Constants file added to make it easier for when interacting with collections from BE.

Affected files:

- ContentActionsButtonViewModel
- WallCell
- UserCell
- ExploreViewModel
- FeedViewModel
- UserContentListViewModel
- ProfileViewModel
- WallFilterViewModel
- WallAppTabView
- WallReplyView
- String
- Constants

---
Notes : A lot of back and forth for pushing this commit - struggled to find a way to efficiently display the reply view all over the app- basically is one view it doesn't matter if opened from profile search or from own profile or wall main view. It will open the ReplyView and I will be working on uploading the replies in the next push.